update: chat context with followers count.
